+++ lib/sinon/spy.js
@@ -7,12 +7,9 @@ var getPropertyDescriptor = require("./util/core/get-property-descriptor");
 var sinonMatch = require("./match");
 var deepEqual = require("./util/core/deep-equal").use(sinonMatch);
 var spyCall = require("./call");
-var timesInWords = require("./util/core/times-in-words");
 var wrapMethod = require("./util/core/wrap-method");
 var sinonFormat = require("./util/core/format");
 var valueToString = require("./util/core/value-to-string");
-var jsDiff = require("diff");
-var chalk = require("chalk");
 
 var push = Array.prototype.push;
 var slice = Array.prototype.slice;
@@ -108,6 +105,8 @@ var uuid = 0;
 
 // Public API
 var spyApi = {
+    formatters: require("./spy-formatters"),
+
     reset: function () {
         if (this.invoking) {
             var err = new Error("Cannot reset Sinon function while invoking it. " +
@@ -414,100 +413,6 @@ delegateToCalls("yieldToOn", false, "yieldToOn", function (property) {
         "' since it was not yet invoked.");
 });
 
-function colorDiffText(diff) {
-    var objects = diff.map(function (part) {
-        var text = part.value;
-        if (part.added) {
-            text = chalk.green(text);
-        } else if (part.removed) {
-            text = chalk.red(text);
-        }
-        if (diff.length === 2) {
-            text += " "; // format simple diffs
-        }
-        return text;
-    });
-    return objects.join("");
-}
-
-function colorSinonMatchText(matcher, calledArg, calledArgMessage) {
-    if (!matcher.test(calledArg)) {
-        matcher.message = chalk.red(matcher.message);
-        if (calledArgMessage) {
-            calledArgMessage = chalk.green(calledArgMessage);
-        }
-    }
-    return calledArgMessage + " " + matcher.message;
-}
-
-spyApi.formatters = {
-    c: function (spyInstance) {
-        return timesInWords(spyInstance.callCount);
-    },
-
-    n: function (spyInstance) {
-        return spyInstance.toString();
-    },
-
-    D: function (spyInstance, args) {
-        var message = "";
-
-        for (var i = 0, l = spyInstance.callCount; i < l; ++i) {
-            // describe multiple calls
-            if (l > 1) {
-                if (i > 0) {
-                    message += "\n";
-                }
-                message += "Call " + (i + 1) + ":";
-            }
-            var calledArgs = spyInstance.getCall(i).args;
-            for (var j = 0; j < calledArgs.length || j < args.length; ++j) {
-                message += "\n";
-                var calledArgMessage = j < calledArgs.length ? sinonFormat(calledArgs[j]) : "";
-                if (sinonMatch.isMatcher(args[j])) {
-                    message += colorSinonMatchText(args[j], calledArgs[j], calledArgMessage);
-                } else {
-                    var expectedArgMessage = j < args.length ? sinonFormat(args[j]) : "";
-                    var diff = jsDiff.diffJson(calledArgMessage, expectedArgMessage);
-                    message += colorDiffText(diff);
-                }
-            }
-        }
-
-        return message;
-    },
-
-    C: function (spyInstance) {
-        var calls = [];
-
-        for (var i = 0, l = spyInstance.callCount; i < l; ++i) {
-            var stringifiedCall = "    " + spyInstance.getCall(i).toString();
-            if (/\n/.test(calls[i - 1])) {
-                stringifiedCall = "\n" + stringifiedCall;
-            }
-            push.call(calls, stringifiedCall);
-        }
-
-        return calls.length > 0 ? "\n" + calls.join("\n") : "";
-    },
-
-    t: function (spyInstance) {
-        var objects = [];
-
-        for (var i = 0, l = spyInstance.callCount; i < l; ++i) {
-            push.call(objects, sinonFormat(spyInstance.thisValues[i]));
-        }
-
-        return objects.join(", ");
-    },
-
-    "*": function (spyInstance, args) {
-        return args.map(sinonFormat).join(", ");
-    }
-};
-
 extend(spy, spyApi);
-
 spy.spyCall = spyCall;
-
 module.exports = spy;
